stupid refactor unsure if salvagable
i had attempted to create this with inhereted structs. I'm not really
sure why other than the idea that it would be mimicing the interfaces.
THis is unecessary as each stuct can be separate and combined into a
super interface. Doing it this way wont affect functionality it will
only affect readability and maintainability.