i've been standardizing the data input with interfaces.
The interfaces should get called to provide data to templates
of generated code. There are two at the moment that i want to g
enerate for. These categories are display and query. One will r
eturn strings for display in templates, and the other will gene
rate the boilerplate for my go framework's sqlc query. currentl
y the incomming data is opposite of what is supposed to happen.
 The model parser generates structs which it will not do i prod
stopping for a break