working great but incomplete
at the moment i am correctly parsing structs and queries and displaying
them for the models.go. unfortunately i didn't realize not all structs
are in the models.go. the query structs are in the .sql.go files.
Luckily my abstraction make it simple to import these files as well. All
i need to do is add the new types.